requirements.txt
Flask
paho-mqtt
requests
